requirements.txt
Pillow=5.2.0